cdk-full/requirements.txt
aws-cdk-lib==2.142.1
constructs>=10.0.0,<11.0.0
fastapi
boto3
mangum